Separate `DataladAuth` (and common helpers) into their own modules
